feat: Added CardsOnHand view to CardDisplaySystem
This one is bit radical but i feel that we can embrace it nicely

- Moved CardsOnHand view composable from ui.components.views to CardDisplaySystem.
- Updated CardsOnHand view to include a badge displaying the number of cards on hand.
- Added scaling to Card composable within CardsOnHand view.
- Adjusted badge positioning in CardsOnHand view.
- Updated preview for CardsOnHand view and added a parameter provider for card entities.
- Renamed DrawCard button to PullCardButton and updated its import in Game composable.
- Updated Game composable to use the new CardsOnHand view and PullCardButton.
- Updated CardDisplaySystem to use componentManager() instead of getTheComponents().
- Added a preview for a random card in CardDisplaySystem.
- Updated CardDisplaySystem to scale the card content.
- Updated componentManager() to return the componentManager instance.
