+++ app/src/main/java/com/sq/thed_ck_licker/ecs/systems/CardDisplaySystem.kt
@@ -5,16 +5,28 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Badge
+import androidx.compose.material3.BadgedBox
+import androidx.compose.material3.Card
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableIntState
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.BiasAlignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.paint
+import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewParameter
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import com.sq.thed_ck_licker.ecs.ComponentManager
 import com.sq.thed_ck_licker.ecs.TheGameHandler
@@ -70,14 +82,52 @@ class CardDisplaySystem(val componentManager: ComponentManager) {
     }
 
 
+    @Composable
+    fun CardsOnHandView(
+        cardsDrawCount: MutableIntState,
+        modifier: Modifier,
+        cardId: Int,
+        cardDisplaySystem: CardDisplaySystem
+    ) {
+
+        BadgedBox(
+            badge = {
+                Badge(
+                    modifier = Modifier.offset((-20).dp, (5).dp),
+                    containerColor = Color.Red
+                ) {
+                    Text("${cardsDrawCount.intValue}")
+                }
+            },
+            modifier = modifier
+                .width(120.dp)
+                .height(170.dp)
+                .background(color = Color.Magenta)
+//            .offset(20.dp, 0.dp)
+//            .scale(0.9f)
+
+        ) {
+            // TODO: There might be some modifier that "just rounds the corners"
+            //  And then it could be just passed via modifier passing or something
+            Card(
+                modifier = modifier
+                    .background(color = Color.Green)
+                    .scale(0.99f)
+            ) {
+                cardDisplaySystem.EntityDisplay(cardId)
+            }
+        }
+    }
+
+
 }
 
 
 @Preview(showBackground = true)
 @Composable
-fun DisplayACardPreview2() {
+fun DisplayRandomCardPreview() {
     TheGameHandler.initTheGame()
-    val displaySystem = CardDisplaySystem(TheGameHandler.getTheComponents())
+    val displaySystem = CardDisplaySystem(TheGameHandler.componentManager())
     Box(
         modifier = Modifier
             .size(500.dp)
@@ -87,3 +137,34 @@ fun DisplayACardPreview2() {
     }
 }
 
+
+
+@Preview(showBackground = true)
+@Composable
+fun CardsOnHandViewPreview(
+    @PreviewParameter(
+        CardEntityPreviewParameterProvider::class,
+        limit = 1
+    ) cardEntity: Int
+) {
+    val displaySystem = CardDisplaySystem(TheGameHandler.componentManager())
+
+    val cardCounter = remember { mutableIntStateOf(69) }
+    displaySystem.CardsOnHandView(
+        cardCounter,
+        Modifier,
+        cardEntity,
+        CardDisplaySystem(TheGameHandler.componentManager())
+    )
+}
+
+
+// TODO: maybe this should be part of test things?
+class CardEntityPreviewParameterProvider : PreviewParameterProvider<Int> {
+    // TODO here should be some way of actually getting card entities
+    init {
+        TheGameHandler.initTheGame()
+    }
+
+    override val values = TheGameHandler.getTestingCardSequence()
+}
